lab5 finish

+++ lab5/JALab5/JALab5/RadToDegAsm.hpp
@@ -1,7 +1,7 @@
 #ifndef RAD_TO_DEG_ASM_HPP
 #define RAD_TO_DEG_ASM_HPP
 
-extern "C" int _stdcall RadToDegAsm(int dRad, bool isSSE2);
+extern "C" double _stdcall RadToDegAsm(int dRad, bool isSSE2);
 
 extern "C" bool _stdcall CpuIdAsm();
